Update README.md
Combining CNNs and LSTMs in hybrid architectures allows the model to simultaneously capture spatial and temporal features, enhancing the overall performance of HAR systems. 

The deep CNN architecture extracts features from an image, which are then processed using the LSTM architecture to generate a caption. 
The CNN-LSTM model is a deep, spatially, and temporally deep architecture used for sequence prediction problems with spatial inputs like images or videos. It combines Convolutional Neural Network (CNN) layers for feature extraction and LSTMs for sequence prediction on feature vectors. This model is at the boundary of Computer Vision and Natural Language Processing, with immense potential for tasks like text classification and video conversion.